Add designSystemDocsDecorator
